fx-charts: move old feature table charts to export package
