+++ mzmine-community/src/main/java/io/github/mzmine/modules/io/export_features_all_speclib_matches/FeatureShapeMobilogramChart.java
@@ -12,7 +12,6 @@
  *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
  * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -23,19 +22,23 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.mzmine.datamodel.features.types.graphicalnodes;
+package io.github.mzmine.modules.io.export_features_all_speclib_matches;
 
 import com.google.common.util.concurrent.AtomicDouble;
-import io.github.mzmine.datamodel.ImagingRawDataFile;
+import io.github.mzmine.datamodel.IMSRawDataFile;
+import io.github.mzmine.datamodel.MobilityType;
 import io.github.mzmine.datamodel.Scan;
+import io.github.mzmine.datamodel.featuredata.IonMobilogramTimeSeries;
 import io.github.mzmine.datamodel.featuredata.IonTimeSeries;
+import io.github.mzmine.datamodel.features.Feature;
 import io.github.mzmine.datamodel.features.ModularFeature;
 import io.github.mzmine.datamodel.features.ModularFeatureListRow;
+import io.github.mzmine.datamodel.features.types.graphicalnodes.BufferedChartNode;
 import io.github.mzmine.datamodel.features.types.modifiers.GraphicalColumType;
 import io.github.mzmine.gui.chartbasics.simplechart.SimpleXYChart;
 import io.github.mzmine.gui.chartbasics.simplechart.datasets.ColoredXYDataset;
 import io.github.mzmine.gui.chartbasics.simplechart.datasets.RunOption;
-import io.github.mzmine.gui.chartbasics.simplechart.providers.impl.series.IonTimeSeriesToXYProvider;
+import io.github.mzmine.gui.chartbasics.simplechart.providers.impl.series.SummedMobilogramXYProvider;
 import io.github.mzmine.gui.preferences.UnitFormat;
 import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.util.RangeUtils;
@@ -45,62 +48,51 @@ import java.util.NoSuchElementException;
 import org.jetbrains.annotations.NotNull;
 import org.jfree.data.Range;
 
-public class FeatureShapeChart extends BufferedChartNode {
+class FeatureShapeMobilogramChart extends BufferedChartNode {
 
-  public FeatureShapeChart(@NotNull ModularFeatureListRow row, AtomicDouble progress) {
+  FeatureShapeMobilogramChart(@NotNull ModularFeatureListRow row, AtomicDouble progress) {
     super(true);
     UnitFormat uf = MZmineCore.getConfiguration().getUnitFormat();
 
-    SimpleXYChart<IonTimeSeriesToXYProvider> chart = new SimpleXYChart<>(
-        uf.format("Retention time", "min"), uf.format("Intensity", "a.u."));
+    final IMSRawDataFile imsFile = (IMSRawDataFile) row.getRawDataFiles().stream()
+        .filter(file -> file instanceof IMSRawDataFile).findAny().orElse(null);
+    if (imsFile == null) {
+      return;
+    }
+    final MobilityType mt = imsFile.getMobilityType();
+    SimpleXYChart<SummedMobilogramXYProvider> chart = new SimpleXYChart<>(mt.getAxisLabel(),
+        uf.format("Intensity", "a.u."));
     chart.setRangeAxisNumberFormatOverride(MZmineCore.getConfiguration().getIntensityFormat());
-    chart.setDomainAxisNumberFormatOverride(MZmineCore.getConfiguration().getRTFormat());
+    chart.setDomainAxisNumberFormatOverride(MZmineCore.getConfiguration().getMobilityFormat());
     chart.setLegendItemsVisible(false);
 
     List<ColoredXYDataset> datasets = new ArrayList<>();
     int size = row.getFilesFeatures().size();
-    for (ModularFeature f : row.getFeatures()) {
-      if (f.getRawDataFile() instanceof ImagingRawDataFile) {
-        continue;
+    for (Feature f : row.getFeatures()) {
+      IonTimeSeries<? extends Scan> series = ((ModularFeature) f).getFeatureData();
+      if (series instanceof IonMobilogramTimeSeries) {
+        datasets.add(new ColoredXYDataset(new SummedMobilogramXYProvider((ModularFeature) f),
+            RunOption.THIS_THREAD));
       }
-      IonTimeSeries<? extends Scan> dpSeries = f.getFeatureData();
-      if (dpSeries != null) {
-        ColoredXYDataset dataset = new ColoredXYDataset(new IonTimeSeriesToXYProvider(f),
-            RunOption.THIS_THREAD);
-        datasets.add(dataset);
-      }
+
       if (progress != null) {
         progress.addAndGet(1.0 / size);
       }
     }
 
     final ModularFeature bestFeature = row.getBestFeature();
-    final org.jfree.data.Range defaultRange;
-    if (bestFeature != null) {
-      final Float rt = bestFeature.getRT();
-
-      var fwhm = bestFeature.getFWHM();
-      var fullWidth = RangeUtils.rangeLength(bestFeature.getRawDataPointsRTRange());
-      var dataRTRange = bestFeature.getRawDataFile().getDataRTRange();
-      var rawMinRt = dataRTRange.lowerEndpoint();
-      var rawMaxRt = dataRTRange.upperEndpoint();
-      // FWHM defines most of the feature / chromatogram
-      if (fwhm != null && !Float.isNaN(fwhm) && fwhm > 0f && fwhm / fullWidth > 0.4) {
-        // zoom on feature
-        var window = 5 * fwhm;
-        defaultRange = new org.jfree.data.Range(Math.max(rt - window, rawMinRt),
-            Math.min(rt + window, rawMaxRt));
-      } else {
-        // show full RT range
-        final float length = Math.max(fullWidth, 0.001f);
-        final double lower = Math.max(rt - length * 1.05, rawMinRt);
-        final double upper = Math.min(rt + length * 1.05, rawMaxRt);
-        // odd possibility for lower > upper in case:
-        // rawMinRt is low and rawMaxRt is high, but rt is 0 and length is small.
-        // (e.g. no data points after msn tree feature list builder)
-        defaultRange = new org.jfree.data.Range(Math.min(lower, upper), Math.max(lower, upper));
+    org.jfree.data.Range defaultRange = null;
+    if (bestFeature != null && bestFeature.getRawDataFile() instanceof IMSRawDataFile imsRaw) {
+      com.google.common.collect.Range<Float> mobilityRange = bestFeature.getMobilityRange();
+      final Float mobility = bestFeature.getMobility();
+      if (mobilityRange != null && mobility != null && !Float.isNaN(mobility)) {
+        final Float length = RangeUtils.rangeLength(mobilityRange);
+        defaultRange = new org.jfree.data.Range(
+            Math.max(mobility - 3 * length, imsRaw.getDataMobilityRange().lowerEndpoint()),
+            Math.min(mobility + 3 * length, imsRaw.getDataMobilityRange().upperEndpoint()));
       }
-    } else {
+    }
+    if (defaultRange == null) {
       defaultRange = new Range(0, 1);
     }
 
@@ -108,12 +100,11 @@ public class FeatureShapeChart extends BufferedChartNode {
     try {
       chart.getXYPlot().getDomainAxis().setRange(defaultRange);
       chart.getXYPlot().getDomainAxis().setDefaultAutoRange(defaultRange);
-    } catch (NullPointerException | NoSuchElementException ex) {
+    } catch (NullPointerException | NoSuchElementException e) {
       // error in jfreechart draw method
     }
 
-    // set the chart to create a buffered image
-    setChartCreateImage(chart, GraphicalColumType.LARGE_GRAPHICAL_CELL_WIDTH,
+    setChartCreateImage(chart, GraphicalColumType.DEFAULT_GRAPHICAL_CELL_WIDTH,
         GraphicalColumType.DEFAULT_GRAPHICAL_CELL_HEIGHT);
   }
 }
